:recycle: makeNewStatesのリファクタリング
